remove the "upgrade-component" and transfert it in spb2aug.component.ts, import.component.html change the target of component for spb2aug.component.ts.
