+++ src/app/components/import/import.component.html
@@ -21,7 +21,7 @@
       <label>
         {{this.multilinguism.translate("import save from snap core first") + ":"}}
       </label>
-      <app-spb2aug-upgrade></app-spb2aug-upgrade>
+      <app-spb2aug></app-spb2aug>
     </div>
 
     <div class="import listElement">
