--- src/app/components/spb2aug-upgrade/spb2aug-upgrade.component.html
@@ -1,1 +1,0 @@
-<input #file type="file" accept=".spb, .sps" (change)="convert(file.files)" onclick="this.value=null">
